requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
djangorestframework==3.12.2
lxml==4.6.2
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
